chore: modularize
