--- src/components/ExperienceList.scss
@@ -1,196 +1,0 @@
-@import '../App.scss';
-
-.experience-list-container {
-  width: 100%;
-  padding: 0;
-  
-  .section-title {
-    font-size: 2rem;
-    font-weight: bold;
-    margin-bottom: 2rem;
-    color: white;
-  }
-  
-  .job-card {
-    width: 100%;
-    padding: 2rem;
-    background-color: #1f2937; // Dark gray background
-    border-radius: 0.75rem;
-    margin-bottom: 1.5rem;
-    transition: opacity 150ms ease;
-    cursor: pointer;
-    
-    &.dimmed {
-      opacity: 0.5;
-    }
-    
-    .job-card-content {
-      display: flex;
-      flex-direction: column;
-      
-      @media (min-width: $mobile-view-transition-threshold) {
-        flex-direction: row;
-      }
-    }
-    
-    .time-frame {
-      margin-bottom: 1rem;
-      
-      @media (min-width: $mobile-view-transition-threshold) {
-        width: 20%;
-        margin-bottom: 0;
-      }
-      
-      p {
-        font-size: 1.125rem;
-        color: #9ca3af; // Gray text
-      }
-    }
-    
-    .job-details {
-      @media (min-width: $mobile-view-transition-threshold) {
-        width: 80%;
-      }
-    }
-    
-    .job-title {
-      display: flex;
-      align-items: center;
-      margin-bottom: 1rem;
-      
-      h2 {
-        font-size: 1.5rem;
-        font-weight: 500;
-        margin-right: 0.5rem;
-        color: white;
-        transition: color 150ms ease;
-        
-        &.hovered {
-          color: #2dd4bf; // Teal color
-        }
-      }
-      
-      svg {
-        transition: color 150ms ease;
-        color: white;
-        
-        &.hovered {
-          color: #2dd4bf; // Teal color
-        }
-      }
-    }
-    
-    .job-description {
-      font-size: 1.125rem;
-      margin-bottom: 1.5rem;
-      color: #d1d5db; // Light gray text
-    }
-    
-    .skill-tags {
-      display: flex;
-      flex-wrap: wrap;
-      gap: 0.5rem;
-      
-      .skill-tag {
-        padding: 0.5rem 1rem;
-        background-color: #374151; // Darker gray
-        border-radius: 9999px; // Full rounded
-        color: #2dd4bf; // Teal text
-        font-size: 0.875rem;
-      }
-    }
-  }
-}
-
-// Expanded card styles
-.overlay-background {
-  position: fixed;
-  inset: 0;
-  background-color: rgba(0, 0, 0, 0.4);
-  transition: background-color 250ms ease;
-  z-index: 40;
-  
-  &.closing {
-    background-color: rgba(0, 0, 0, 0);
-  }
-}
-
-.expanded-card {
-  color: #d1d5db; // Light gray text
-  
-  .back-button-container {
-    transition: opacity 150ms ease;
-  }
-  
-  .back-button {
-    display: flex;
-    align-items: center;
-    background: none;
-    border: none;
-    color: white;
-    cursor: pointer;
-    padding: 0;
-    margin-bottom: 1.5rem;
-    transition: color 150ms ease;
-    
-    &:hover {
-      color: #2dd4bf; // Teal on hover
-    }
-    
-    span {
-      margin-left: 0.25rem;
-    }
-  }
-  
-  .expanded-title {
-    font-size: 1.875rem;
-    font-weight: 500;
-    margin-bottom: 0.5rem;
-    color: #2dd4bf; // Teal color
-  }
-  
-  .expanded-timerange {
-    font-size: 1.125rem;
-    margin-bottom: 1.5rem;
-    color: #9ca3af; // Gray text
-  }
-  
-  .expanded-main-content {
-    margin-bottom: 2rem;
-    
-    p {
-      font-size: 1.125rem;
-      margin-bottom: 1.5rem;
-    }
-  }
-  
-  .expanded-details {
-    margin-top: 2.5rem;
-    border-top: 1px solid #4b5563; // Gray border
-    padding-top: 1.5rem;
-    
-    &.hidden {
-      display: none;
-    }
-  }
-  
-  .detail-section {
-    margin-bottom: 1.5rem;
-    
-    h3 {
-      font-size: 1.25rem;
-      margin-bottom: 0.75rem;
-      color: white;
-    }
-    
-    ul {
-      list-style-type: disc;
-      padding-left: 1.25rem;
-      
-      li {
-        color: #d1d5db; // Light gray text
-        margin-bottom: 0.5rem;
-      }
-    }
-  }
-}
