+++ src/components/experience/JobTitle.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { ArrowUpRight } from "lucide-react";
+import { JobTitleProps } from "../../types/experience";
+
+const JobTitle: React.FC<JobTitleProps> = ({ title, company, isHovered }) => (
+    <div className="job-title">
+        <h2 className={isHovered ? "hovered" : ""}>
+            {title} · {company}
+        </h2>
+        <ArrowUpRight className={isHovered ? "hovered" : ""} size={20} />
+    </div>
+);
+
+export default JobTitle;
