+++ src/components/experience/index.ts
@@ -1,0 +1,14 @@
+import ExperienceList from "./ExperienceList";
+
+export default ExperienceList;
+
+// Export individual components for reuse elsewhere if needed
+export { default as TimeFrame } from "./TimeFrame";
+export { default as JobTitle } from "./JobTitle";
+export { default as JobDescription } from "./JobDescription";
+export { default as SkillTag } from "./SkillTag";
+export { default as SkillTags } from "./SkillTags";
+export { default as DetailSection } from "./DetailSection";
+export { default as BackButton } from "./BackButton";
+export { default as ExpandedCardOverlay } from "./ExpandedCardOverlay";
+export { default as JobCard } from "./JobCard";
