+++ src/types/experience.ts
@@ -1,0 +1,72 @@
+import { ReactNode } from "react";
+
+export interface TimeFrameProps {
+    timeRange: string;
+}
+
+export interface JobTitleProps {
+    title: string;
+    company: string;
+    isHovered: boolean;
+}
+
+export interface JobDescriptionProps {
+    description: string;
+}
+
+export interface SkillTagProps {
+    skill: string;
+}
+
+export interface SkillTagsProps {
+    skills: string[];
+}
+
+export interface DetailSectionProps {
+    title: string;
+    items: string[];
+}
+
+export interface BackButtonProps {
+    onClick: () => void;
+}
+
+export interface ExpandedCardOverlayProps {
+    job: JobExperience;
+    rect: DOMRect;
+    mainContainerRect: DOMRect;
+    onClose: () => void;
+    isAnimating: boolean;
+    isClosing: boolean;
+}
+
+export interface JobCardProps {
+    id: number;
+    timeRange: string;
+    title: string;
+    company: string;
+    description: string;
+    skills: string[];
+    isHovered: boolean;
+    onMouseEnter: (id: number) => void;
+    onMouseLeave: () => void;
+    onClick: (id: number, rect: DOMRect) => void;
+    hoveredId: number | null;
+    expandedId: number | null;
+}
+
+export interface JobDetails {
+    achievements: string[];
+    projects: string[];
+    technologies: string[];
+}
+
+export interface JobExperience {
+    id: number;
+    timeRange: string;
+    title: string;
+    company: string;
+    description: string;
+    skills: string[];
+    details: JobDetails;
+}
